Fix assemblies only publicized when Type publicized
An assembly is currently marked as publicized only if one of its Types changes from non-public to public. This means that specificly publicizing e.g. a non-public field in an already-public Type isn't marked as a modified assembly, in which case no publicized assembly is saved to disk.
This fix makes it so that any member being publicized marks the assembly as modified and saved to disk.